requirements.txt
corner==2.2.1
decorator>=4.2.1
dill>=0.3.1.1
dynesty==2.1.2
typing-inspect>=0.4.0
emcee>=3.1.3
matplotlib
numpydoc>=1.0.0
pyprojroot==0.2.0
pyswarms==1.3.0
h5py>=2.10.0
SQLAlchemy==1.3.20
scipy<=1.11.3
astunparse==1.6.3
threadpoolctl>=3.1.0,<=3.2.0
timeout-decorator==0.5.0
xxhash<=3.4.1
